Replaced TailwindCSS with plain CSS to resolve continuous deployment issues on Amplify
