+++ package.json
@@ -20,10 +20,9 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
-    "autoprefixer": "^10.4.21",
     "eslint": "^9",
-    "eslint-config-next": "15.1.8",
-    "postcss": "^8.5.3",
-    "tailwindcss": "^3.4.17"
+    "eslint-config-next": "15.1.8"
   }
 }
+
+
