+++ src/app/page.jsx
@@ -1,9 +1,9 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import Image from "next/image";
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import Navbar from '@/components/Navbar'; 
+import Navbar from '@/components/Navbar/Navbar'; 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import Footer from '@/components/Footer';
+import Footer from '@/components/Footer/Footer';
 
 export default function Home() {
   return (
