+++ src/styles/global.css
@@ -1,0 +1,179 @@
+/* /src/styles/utility.css */
+
+/* Spacing */
+.padding-x {
+  padding-left: 1.5rem;
+  padding-right: 1.5rem;
+}
+
+@media (min-width: 640px) {
+  .padding-x {
+    padding-left: 4rem;
+    padding-right: 4rem;
+  }
+}
+
+.padding-y {
+  padding-top: 1.5rem;
+  padding-bottom: 1.5rem;
+}
+
+@media (min-width: 640px) {
+  .padding-y {
+    padding-top: 4rem;
+    padding-bottom: 4rem;
+  }
+}
+
+.padding {
+  padding: 2.5rem 1.5rem;
+}
+
+@media (min-width: 640px) {
+  .padding {
+    padding: 4rem;
+  }
+}
+
+/* Text sizes */
+.hero-head-text {
+  font-weight: 900;
+  color: white;
+  font-size: 40px;
+  line-height: normal;
+  margin-top: 0.5rem;
+}
+
+@media (min-width: 640px) {
+  .hero-head-text {
+    font-size: 60px;
+  }
+}
+@media (min-width: 1024px) {
+  .hero-head-text {
+    font-size: 80px;
+    line-height: 98px;
+  }
+}
+
+.hero-sub-text {
+  font-weight: 500;
+  color: #dfd9ff;
+  font-size: 16px;
+}
+
+@media (min-width: 640px) {
+  .hero-sub-text {
+    font-size: 26px;
+  }
+}
+@media (min-width: 1024px) {
+  .hero-sub-text {
+    font-size: 30px;
+    line-height: 40px;
+  }
+}
+
+/* Section headings */
+.section-head-text {
+  color: white;
+  font-weight: 900;
+  font-size: 30px;
+  transition: all 1s ease-in-out;
+}
+
+@media (min-width: 640px) {
+  .section-head-text {
+    font-size: 40px;
+  }
+}
+@media (min-width: 768px) {
+  .section-head-text {
+    font-size: 60px;
+  }
+}
+
+.section-sub-text {
+  text-transform: uppercase;
+  font-size: 14px;
+  letter-spacing: 0.1em;
+  color: var(--color-secondary); /* define this in :root */
+}
+
+@media (min-width: 640px) {
+  .section-sub-text {
+    font-size: 18px;
+  }
+}
+
+/* Footer */
+.footer {
+  background-color: var(--color-primary); /* define this in :root */
+  color: white;
+  padding: 2rem 1rem;
+}
+
+.footer-container {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  max-width: 1120px;
+  margin: 0 auto;
+}
+
+.footer-brand {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+}
+
+.footer-logo {
+  width: 5rem;
+  height: auto;
+  margin-bottom: 1rem;
+}
+
+.footer-text {
+  font-size: 0.875rem;
+  color: white;
+}
+
+.footer-links {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  gap: 1rem;
+}
+
+.link-column {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.link-title {
+  font-weight: 600;
+  font-size: 1.125rem;
+  margin-bottom: 0.5rem;
+}
+
+.link-item {
+  font-size: 0.875rem;
+  color: #d1d5db;
+  transition: color 0.3s;
+}
+
+.link-item:hover {
+  color: white;
+}
+
+.icon {
+  font-size: 2rem;
+  color: #d1d5db;
+  transition: color 0.3s;
+}
+
+.icon:hover {
+  color: white;
+}
